# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: API – Sistema Web de Votaciones del Festival de la Leyenda Vallenata
  version: 1.0.0
  description: 'API REST del Sistema Web de Votaciones del Festival de la Leyenda
    Vallenata (proyecto académico, Areandina – Desarrollo Web). Jerarquía: Edición
    → Categoría → Votación → Opción → Voto. Autenticación: encabezado `Authorization:
    Token <token>` obtenido en `/api/auth/login/` o `/api/auth/registro/`.'
paths:
...
  /api/admin/votaciones/{id}/:
    get:
      operationId: admin_votaciones_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este votación.
        required: true
      tags:
      - Administración
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Votacion'
          description: ''
    put:
      operationId: admin_votaciones_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este votación.
        required: true
      tags:
      - Administración
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VotacionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VotacionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VotacionRequest'
        required: true
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Votacion'
          description: ''
    patch:
      operationId: admin_votaciones_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este votación.
        required: true
      tags:
      - Administración
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedVotacionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedVotacionRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedVotacionRequest'
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Votacion'
          description: ''
    delete:
      operationId: admin_votaciones_destroy
      summary: Eliminar (solo sin votos, RN-09)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este votación.
        required: true
      tags:
      - Administración
      security:
      - tokenAuth: []
      responses:
        '204':
          description: No response body
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRegla'
          description: ''
...
  /api/admin/votaciones/{id}/publicar/:
    post:
      operationId: admin_votaciones_publicar_create
      summary: Publicar votación (RN-06)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este votación.
        required: true
      tags:
      - Administración
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Votacion'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRegla'
          description: ''
  /api/admin/votaciones/{id}/publicar-resultados/:
    post:
      operationId: admin_votaciones_publicar_resultados_create
      summary: Publicar u ocultar resultados (RF-15)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este votación.
        required: true
      tags:
      - Administración
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicarResultadosRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PublicarResultadosRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PublicarResultadosRequest'
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Votacion'
          description: ''
...
  /api/admin/votaciones/{id}/resultados/csv/:
    get:
      operationId: admin_votaciones_resultados_csv_retrieve
      summary: Exportar resultados en CSV (RF-14)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este votación.
        required: true
      tags:
      - Administración
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            text/csv:
              schema:
                type: string
                format: binary
          description: Archivo CSV
...
  /api/votaciones/{id}/:
    get:
      operationId: votaciones_retrieve
      summary: Detalle con opciones y votos del usuario (RF-06)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este votación.
        required: true
      tags:
      - Consulta pública
      security:
      - tokenAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VotacionDetalle'
          description: ''
...
  /api/votaciones/por-ruta/:
    get:
      operationId: votaciones_por_ruta_retrieve
      summary: Detalle por URL amigable /{año}/{categoría}/{votación} (RF-06)
      parameters:
      - in: query
        name: anio
        schema:
          type: integer
        description: Año de la edición
        required: true
      - in: query
        name: categoria
        schema:
          type: string
        description: Slug de la categoría
        required: true
      - in: query
        name: votacion
        schema:
          type: string
        description: Slug de la votación
        required: true
      tags:
      - Consulta pública
      security:
      - tokenAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VotacionDetalle'
          description: ''
        '404':
          description: No existe o no está publicada
components:
  schemas:
...
    ErrorRegla:
      type: object
      properties:
        detail:
          type: string
          description: Mensaje legible; cita la regla de negocio (RN-xx).
        codigo:
          type: string
          description: votacion_no_abierta · opcion_invalida · limite_votos · opciones_insuficientes
            · tiene_votos · registros_asociados
      required:
      - codigo
      - detail
...
    PatchedVotacionRequest:
      type: object
      description: El slug es opcional (se genera desde el nombre o título), así que
        su unicidad se valida aparte.
      properties:
        categoria:
          type: integer
        titulo:
          type: string
          minLength: 1
          maxLength: 150
        slug:
          description: Identificador para la URL pública; se genera desde el título
            si se deja vacío.
          oneOf:
          - type: string
            maxLength: 150
            pattern: ^[-a-zA-Z0-9_]+$
          - type: string
            maxLength: 0
        descripcion:
          type: string
        imagen:
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        fecha_apertura:
          type: string
          format: date-time
        fecha_cierre:
          type: string
          format: date-time
        votos_por_usuario:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        visibilidad_resultados:
          $ref: '#/components/schemas/VisibilidadResultadosEnum'
    PublicarResultadosRequest:
      type: object
      properties:
        publicar:
          type: boolean
          default: true
...
    VisibilidadResultadosEnum:
      enum:
      - tiempo_real
      - al_cierre
      - no_publicar
      type: string
      description: |-
        * `tiempo_real` - En tiempo real
        * `al_cierre` - Al cierre
        * `no_publicar` - No publicar
    Votacion:
      type: object
      description: El slug es opcional (se genera desde el nombre o título), así que
        su unicidad se valida aparte.
      properties:
        id:
          type: integer
          readOnly: true
        categoria:
          type: integer
        categoria_slug:
          type: string
          readOnly: true
        edicion_anio:
          type: integer
          readOnly: true
        titulo:
          type: string
          maxLength: 150
        slug:
          description: Identificador para la URL pública; se genera desde el título
            si se deja vacío.
          oneOf:
          - type: string
            maxLength: 150
            pattern: ^[-a-zA-Z0-9_]+$
          - type: string
            maxLength: 0
        descripcion:
          type: string
        imagen:
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        fecha_apertura:
          type: string
          format: date-time
        fecha_cierre:
          type: string
          format: date-time
        votos_por_usuario:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        visibilidad_resultados:
          $ref: '#/components/schemas/VisibilidadResultadosEnum'
        estado:
          type: string
          readOnly: true
        publicada:
          type: boolean
          readOnly: true
        cerrada_manualmente:
          type: boolean
          readOnly: true
        resultados_publicados:
          type: boolean
          readOnly: true
        creada_en:
          type: string
          format: date-time
          readOnly: true
        actualizada_en:
          type: string
          format: date-time
          readOnly: true
      required:
      - actualizada_en
      - categoria
      - categoria_slug
      - cerrada_manualmente
      - creada_en
      - edicion_anio
      - estado
      - fecha_apertura
      - fecha_cierre
      - id
      - publicada
      - resultados_publicados
      - titulo
...
    VotacionRequest:
      type: object
      description: El slug es opcional (se genera desde el nombre o título), así que
        su unicidad se valida aparte.
      properties:
        categoria:
          type: integer
        titulo:
          type: string
          minLength: 1
          maxLength: 150
        slug:
          description: Identificador para la URL pública; se genera desde el título
            si se deja vacío.
          oneOf:
          - type: string
            maxLength: 150
            pattern: ^[-a-zA-Z0-9_]+$
          - type: string
            maxLength: 0
        descripcion:
          type: string
        imagen:
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        fecha_apertura:
          type: string
          format: date-time
        fecha_cierre:
          type: string
          format: date-time
        votos_por_usuario:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        visibilidad_resultados:
          $ref: '#/components/schemas/VisibilidadResultadosEnum'
      required:
      - categoria
      - fecha_apertura
      - fecha_cierre
      - titulo
...
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"
...